requirements.txt
anthropic>=0.49.0
openai>=1.0.0
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24
pyautogui>=0.9.54
pystray>=0.19.5
pynput>=1.7
pillow>=10.0
